third_party/WebKit/LayoutTests/fast/forms/select/popup-menu-mouse-operations.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
<script src="../resources/common.js"></script>
<script src="../resources/picker-common.js"></script>
</head>
<body>
<select id="menu">
  <option>foo</option>
  <option selected>bar</option>
  <option>baz</option>
</select>
<script>
var menuElement = document.getElementById('menu');
var picker = null;
var clickEventCounter = 0;
menuElement.addEventListener('click', function() { clickEventCounter++; }, false);
var mouseupEventCounter = 0;
menuElement.addEventListener('mouseup', function() { mouseupEventCounter++; }, false);

function openPickerErrorCallback() {
    testFailed('picker didn\'t open')
    finishJSTest();
}
openPicker(menu, test1, openPickerErrorCallback);
function test1() {
    picker = window.internals.pagePopupWindow.global.picker;
    shouldBeEqualToString('picker._selectElement.value', '1');
    shouldBeEqualToString('menuElement.value', 'bar');

    hoverOverElement(picker._selectElement.children[0]);
    shouldBeEqualToString('picker._selectElement.value', '0');
    shouldBeEqualToString('menuElement.value', 'bar');

    hoverOverElement(picker._selectElement.children[1]);
    shouldBeEqualToString('picker._selectElement.value', '1');
    shouldBeEqualToString('menuElement.value', 'bar');

    hoverOverElement(menuElement);
    shouldBeEqualToString('picker._selectElement.value', '1');
    shouldBeEqualToString('menuElement.value', 'bar');

    // Start drag selecting but release outside the popup.
    hoverOverElement(picker._selectElement.children[2]);
    eventSender.mouseDown();
    hoverOverElement(picker._selectElement.children[0]);
    hoverOverElement(menuElement);
    eventSender.mouseUp();
    shouldNotBe('window.internals.pagePopupWindow', 'null');
    shouldBeEqualToString('picker._selectElement.value', '0');
    shouldBeEqualToString('menuElement.value', 'bar');

    shouldBe('clickEventCounter', '0');
    shouldBe('mouseupEventCounter', '0');

    clickElement(picker._selectElement.children[2]);
    shouldBeNull('window.internals.pagePopupWindow');
    shouldBeEqualToString('menuElement.value', 'baz');

    shouldBe('clickEventCounter', '1');
    shouldBe('mouseupEventCounter', '1');

    waitUntilClosing(function() {
        openPicker(menu, test2, openPickerErrorCallback);
    });
}

function test2() {
    picker = window.internals.pagePopupWindow.global.picker;
    shouldBeEqualToString('picker._selectElement.value', '2');
    shouldBeEqualToString('internals.selectMenuListText(menu)', 'baz');

    eventSender.keyDown('upArrow');
    shouldBeEqualToString('picker._selectElement.value', '1');
    shouldBeEqualToString('menu.value', 'baz');
    shouldBeEqualToString('internals.selectMenuListText(menu)', 'bar');

    // click outside to close popup
    eventSender.mouseMoveTo(300, 1);
    eventSender.mouseDown();
    eventSender.mouseUp();
    shouldBeNull('window.internals.pagePopupWindow');
    shouldBeEqualToString('menu.value', 'bar');
    shouldBeEqualToString('internals.selectMenuListText(menu)', 'bar');

    finishJSTest();
}


</script>
</body>
</html>
